// src/models/Steam.ts
export interface Game {
  appid: number;
  name: string;
  playtime_forever: number;
  img_icon_url: string;
  img_logo_url: string;
  has_community_visible_stats: boolean;
  playtime_windows_forever: number;
  playtime_mac_forever: number;
  playtime_linux_forever: number;
  playtime_2weeks?: number;
}

export default interface Steam {
  game_count: number;
  games: Game[];
}